fixed NPE, fixed TimeSeenMS being parsed as int rather then long resulting in negative integer, Added list commands. QA testing required.

+++ src/main/java/net/dirtcraft/ftbintegration/utility/NbtHelper.java
@@ -8,7 +8,7 @@ public class NbtHelper {
         else return def;
     }
     public static long getOrDefault(NBTTagCompound nbt, String key, long def) {
-        if (nbt.hasKey(key)) return nbt.getInteger(key);
+        if (nbt.hasKey(key)) return nbt.getLong(key);
         else return def;
     }
 }
